Util: Update Format::Builder floating-point precision

+++ src/util/format/builder.h
@@ -7,7 +7,7 @@
 #pragma once
 
 #include <cstddef> // size_t
-#include <cstdint> // int32_t, uint32_t, int64_t
+#include <cstdint> // int32_t, int64_t, uint8_t, uint32_t
 #include <sstream> // stringstream
 #include <string_view>
 
@@ -26,8 +26,8 @@ public:
 	void putU32(uint32_t number) const { m_builder << number; } // unsigned int
 	void putI64(int64_t number) const { m_builder << number; }  // long int
 	void putU64(size_t number) const { m_builder << number; }   // long unsigned int
-	void putF32(float number, size_t precision = 6) const;
-	void putF64(double number, size_t precision = 6) const;
+	void putF32(float number, uint8_t precision = 6) const;
+	void putF64(double number, uint8_t precision = 6) const;
 	void putCharacter(char character) const { m_builder.write(&character, 1); }
 	void putString(const std::string_view string) const { m_builder.write(string.data(), string.length()); }
 	void putPointer(const void* pointer) const { m_builder << pointer; }
